requirements.txt
fastapi
uvicorn
aiosqlite
openai
pydantic
python-dotenv
httpx
